back/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.4.0
colorama==0.4.6
colorthief==0.2.1
constantly==23.10.4
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
daphne==4.1.2
decorator==5.1.1
Django==4.2.16
django-cors-headers==4.6.0
django-environ
django-extensions==3.2.3
djangorestframework==3.15.2
executing==2.1.0
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
hyperlink==21.0.0
idna==3.10
importlib_resources==6.4.5
incremental==24.7.2
ipython
jedi==0.19.2
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.2.1
pypalette==0.1.6
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
service-identity==24.2.0
setuptools==75.5.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.1
tomli==2.0.2
torchsummary==1.5.1
tqdm==4.66.5
traitlets==5.14.3
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.21.0
zope.interface==7.1.1
